+++ gzap/apworld/uzdoom/model/DoomLocation.py
@@ -75,7 +75,7 @@ class DoomLocation(DoomReachable):
         return self.categories & frozenset(args)
 
     def is_default_enabled(self) -> bool:
-        if self.has_category('weapon', 'key', 'token', 'ap_flag', 'big'):
+        if self.has_category('ap_progression', 'ap_useful', 'big'):
             return True
         if self.has_category('medium', 'small', 'tool'):
             return False